refactor(governance): drop admission_proof from the readiness beacon
Direct admission replaced the path that read this field, and #3732 deleted
its last reader. What was left was a `SignedGroupOpenInvitation` stapled to
every beacon a joining node emitted, signed as part of the beacon body and
read by nobody.

Removing it restores the pre-proof beacon layout exactly: seven fields, no
Option tag. That is still a wire break in both directions - the struct's
shape changed for every beacon, not only proof-carrying ones - so a node on
either shape drops every beacon from a node on the other until a fleet
finishes upgrading. Nothing is persisted and no state is corrupted; beacon
liveness and divergence detection go dark between a mismatched pair and
self-heal on completion.

Tests follow the field:

- `beacon_signature_covers_admission_proof` retires with what it covered.
- `old_beacon_layout_rejects_new_encoding` asserted that a beacon fails to
  decode as the seven-field shape - which is now the shape it has. Re-aimed
  as `proof_carrying_layout_rejects_new_encoding`, pinning the same
  fail-closed contract for the break this change introduces.
- `beacon_round_trips` re-pins the byte length without the Option tag.
- The two emit-path tests and `unprovable_beacon_pulls_nothing` go, per the
  note left on them when their behaviour was deleted.

`provable_beacon_e2e` keeps one test that never depended on the proof, so it
is renamed `membership_republish_e2e` for what it now covers. Removing the
last readers of `TestNode::readiness_cache` and `TestNode::stream_opens`
made both dead; they go too, along with two comments left describing the
"provable arm" that #3732 removed.
