+++ crates/governance-types/AGENTS.md
@@ -42,7 +42,7 @@ cargo test -p calimero-governance-types sign_and_verify_round_trip -- --nocaptur
 | `GovernanceError` | enum (`thiserror`) | `SchemaVersion`, `Signature` (from `SignatureError`), `BorshSerialize` (from `io::Error`), `Bounds` |
 | `bounds` (module) | consts | Anti-amplification caps for decoded-op field lengths (see Invariants) |
 | `wire::NamespaceTopicMsg` / `wire::GroupTopicMsg` | enum | Discriminated gossipsub envelope: `Op`, `Ack`, `ReadinessBeacon`, `ReadinessProbe`, (+`MigrationHeartbeat` on the namespace topic) |
-| `wire::SignedAck`, `wire::SignedReadinessBeacon`, `wire::SignedMigrationHeartbeat`, `wire::ReadinessProbe` | struct | Signed gossip primitives layered on top of ops (see Mental Model); `SignedReadinessBeacon` additionally carries an `admission_proof: Option<SignedGroupOpenInvitation>` field, positioned immediately before `signature` and covered by the beacon signature |
+| `wire::SignedAck`, `wire::SignedReadinessBeacon`, `wire::SignedMigrationHeartbeat`, `wire::ReadinessProbe` | struct | Signed gossip primitives layered on top of ops (see Mental Model) |
 
 ## Mental Model
 
@@ -78,7 +78,7 @@ The gossip layer (`wire.rs`) is a separate concern layered on top of ops: `Names
 - **`ContextCapabilityBits` cannot be zero, enforced twice**: `new()`/`try_from` reject it at construction, and a hand-written `BorshDeserialize` impl rejects a zero byte on the wire too - so a zero-capability op cannot even be constructed by decoding untrusted bytes. Unknown non-zero bits are accepted (forward-compatible replay by older nodes); callers must mask against known constants before interpreting.
 - **`bounds` module and `validate()` exist purely against a malicious/buggy peer.** Legitimate ops are already capped by `MAX_SIGNED_GROUP_OP_PAYLOAD_BYTES` (64 KiB) upstream on the send path; borsh itself has no inherent per-element cap when decoding gossip/backfill bytes, so `SignedGroupOp::validate()`/`SignedNamespaceOp::validate()` reject egregiously oversized decoded ops (parent list, ciphertext, key-envelope count, id lists, TEE allow-list entries, metadata maps) before they reach apply/storage. Call `validate()` right after decode, before doing anything else with an untrusted op.
 - **`SignedMigrationHeartbeat` has a hand-rolled `BorshDeserialize`, not the derive.** `authored_remaining` (u64) and `migration_failed` (u8) are trailing fields appended AFTER the signed body, read via `read_trailing` with EOF-tolerance so an old-format heartbeat missing them decodes to `0` rather than erroring. Both fields are deliberately OUTSIDE the signed payload (advisory telemetry, not a migration gate) - tampering them never breaks `verify_signature()`. Adding a new field: append after `migration_failed` (another trailing read) or introduce a version discriminant; do NOT reorder the existing prefix fields, and keep the round-trip + mixed-fleet tests in `wire.rs` in step.
-- **`SignedReadinessBeacon`'s `admission_proof` field did NOT follow the `SignedMigrationHeartbeat` precedent above, and that gap is unresolved, not overlooked.** The field is a plain derived-borsh struct member, signed as part of the body, not a hand-rolled trailing read - so unlike the heartbeat, this crate's usual mixed-fleet tolerance was not applied here. `NamespaceTopicMsg`/`GroupTopicMsg` decode via `borsh::from_slice`, which rejects trailing bytes, so an un-upgraded node fails closed on any new-format beacon. Because the struct's shape changed for every beacon, not only proof-carrying ones, this drops EVERY beacon an upgraded node emits, `admission_proof: None` included. Beacon-driven liveness and divergence detection go dark between a mismatched pair until the rollout completes, then self-heal; nothing is persisted and no state is corrupted, but the change requires a coordinated cluster upgrade.
+- **`SignedReadinessBeacon` carries no optional fields, and a new one must follow the `SignedMigrationHeartbeat` precedent above.** The beacon previously carried an `admission_proof: Option<SignedGroupOpenInvitation>` as a plain derived-borsh member signed as part of the body - not a hand-rolled trailing read - so this crate's usual mixed-fleet tolerance never applied to it. That field is gone with the direct-admission cutover, which restores the pre-proof layout exactly. Both the addition and the removal changed the struct's shape for EVERY beacon, not only proof-carrying ones: `NamespaceTopicMsg`/`GroupTopicMsg` decode via `borsh::from_slice`, which rejects trailing bytes, so a mismatched node fails closed and drops every beacon its peer emits. Beacon-driven liveness and divergence detection go dark between a mismatched pair until the rollout completes, then self-heal; nothing is persisted and no state is corrupted, but either direction requires a coordinated cluster upgrade. `proof_carrying_layout_rejects_new_encoding` in `wire.rs` pins the fail-closed half of that contract.
 - **`verify_signature()` is integrity, not authorization.** Every `verify_signature()` doc comment says this explicitly - treating a passing verify as "this signer is allowed to do this" is a real vulnerability class (accepting ops from revoked/never-admitted keys). Authorization is the caller's job.
 - **`RootOp::NamespaceCreated` is the self-authorizing namespace genesis** (#2474): it is the only op whose apply does NOT call `require_namespace_admin`, because it's what establishes admin authority in the first place. Its anti-hijack property (apply is a no-op if root meta already exists) only protects an already-established namespace - the very first genesis on a bare replica is trust-on-first-sync (see #2932 in the doc comment).
 - Domain separation prefixes (`GROUP_GOVERNANCE_SIGN_DOMAIN`, `NAMESPACE_GOVERNANCE_SIGN_DOMAIN`, `KEY_ENVELOPE_SIGN_DOMAIN`, `wire::ACK_SIGN_DOMAIN`, `wire::READINESS_BEACON_SIGN_DOMAIN`, `wire::MIGRATION_HEARTBEAT_SIGN_DOMAIN`) are each distinct so a signature can never be lifted from one protocol surface (e.g. an ack) and replayed as another (e.g. an op). Never reuse a domain string across two signable-bytes functions.
