+++ crates/node/src/lib.rs
@@ -67,7 +67,7 @@ mod parent_plan_applied_flag_test;
 // Also on the ungated harness: no mock attestation, runs in the default
 // `cargo test`.
 #[cfg(test)]
-mod provable_beacon_e2e;
+mod membership_republish_e2e;
 /// Feature-ungated (`#[cfg(test)]`-only) in-process node harness shared by
 /// `local_governance_node_e2e` and `cascade_dispatch_e2e`. Contains no mock
 /// attestation code — see the module docs before adding any.
